requirements.txt
Mastodon.py
pylast
python-dotenv